requirements.txt
python-dotenv==1.0.0
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.3
transformers==4.36.2
torch==2.1.2
TTS==0.22.0
requests==2.31.0
webrtcvad==2.0.10
pydub==0.25.1
soundfile==0.12.1 